chore(eslint-plugin): refactor explicit return type rules to share code
I saw this duplication, but let it go in the PR because it was a non-trivial refactor.
This PR just cleans up that tech debt.
